1.点位导入优化 2.模型同步优化

+++ cloud_api_serve_git/sample/src/main/java/com/dji/sample/center/entity/UniPoint.java
@@ -129,7 +129,7 @@ public class UniPoint {
      * <28>: = 避雷器动作次数表
      */
     @TableField("device_type")
-    private Long deviceType;
+    private Integer deviceType;
 
     /**
      * 表计类型
